+++ scripts/__init__.py
@@ -1,30 +1,1 @@
-"""
-Test plan artifact utilities.
-
-Provides schema validation, frontmatter read/write, and artifact management
-for test plans, test cases, and test gaps.
-"""
-
-from .artifact_utils import (
-    SCHEMAS,
-    validate,
-    apply_defaults,
-    detect_schema_type,
-    get_schema_yaml,
-    read_frontmatter,
-    read_frontmatter_validated,
-    write_frontmatter,
-    update_frontmatter,
-)
-
-__all__ = [
-    "SCHEMAS",
-    "validate",
-    "apply_defaults",
-    "detect_schema_type",
-    "get_schema_yaml",
-    "read_frontmatter",
-    "read_frontmatter_validated",
-    "write_frontmatter",
-    "update_frontmatter",
-]
+"""Test plan scripts."""
