+++ scripts/filter_for_revision.py
@@ -18,7 +18,7 @@ import os
 import sys
 
 sys.path.insert(0, os.path.dirname(os.path.abspath(__file__)))
-from artifact_utils import read_frontmatter_validated, update_frontmatter
+from scripts.utils.frontmatter_utils import read_frontmatter_validated, update_frontmatter
 
 
 def main():
