+++ scripts/utils/frontmatter_utils.py
@@ -1,0 +1,145 @@
+"""
+Frontmatter read/write/validate operations for test plan artifacts.
+
+Handles YAML frontmatter in markdown files (between --- delimiters).
+Uses schemas.py for validation.
+"""
+
+import re
+import sys
+
+try:
+    import yaml
+except ImportError:
+    print(
+        "Error: PyYAML is required but not installed.\n"
+        "Install it with: uv pip install -r requirements.txt",
+        file=sys.stderr,
+    )
+    sys.exit(1)
+
+from .schemas import ValidationError, validate, apply_defaults
+
+
+# ─── Frontmatter Read/Write ────────────────────────────────────────────────────
+
+_FRONTMATTER_RE = re.compile(
+    r'^---\s*\n(.*?\n)---\s*\n', re.DOTALL)
+
+
+def read_frontmatter(path):
+    """Read and parse YAML frontmatter from a markdown file.
+
+    Returns:
+        (data_dict, body_string) — frontmatter as dict, remainder as string.
+        Returns ({}, full_content) if no frontmatter found.
+    """
+    with open(path, encoding="utf-8") as f:
+        content = f.read()
+
+    match = _FRONTMATTER_RE.match(content)
+    if not match:
+        return {}, content
+
+    yaml_str = match.group(1)
+    body = content[match.end():]
+
+    data = yaml.safe_load(yaml_str)
+    if not isinstance(data, dict):
+        return {}, content
+
+    return data, body
+
+
+def read_frontmatter_validated(path, schema_type):
+    """Read frontmatter and validate against schema.
+
+    Returns:
+        (data_dict, body_string)
+
+    Raises:
+        ValidationError: if frontmatter fails validation
+        FileNotFoundError: if file doesn't exist
+    """
+    data, body = read_frontmatter(path)
+    if not data:
+        raise ValidationError(f"No frontmatter found in {path}")
+
+    errors = validate(data, schema_type)
+    if errors:
+        raise ValidationError(
+            f"Frontmatter validation failed in {path}:\n"
+            + "\n".join(f"  - {e}" for e in errors))
+
+    return data, body
+
+
+def write_frontmatter(path, data, schema_type):
+    """Write frontmatter to a markdown file.
+
+    Validates data against schema, writes frontmatter + body.
+
+    Args:
+        path: file path (must exist if preserving body, or new file)
+        data: frontmatter dict
+        schema_type: schema to validate against
+
+    Raises:
+        ValidationError: if data fails validation
+    """
+    apply_defaults(data, schema_type)
+    errors = validate(data, schema_type)
+    if errors:
+        raise ValidationError(
+            f"Frontmatter validation failed:\n"
+            + "\n".join(f"  - {e}" for e in errors))
+
+    # Try to preserve existing body
+    try:
+        _, body = read_frontmatter(path)
+    except FileNotFoundError:
+        body = ""
+
+    yaml_str = yaml.dump(data, default_flow_style=False, sort_keys=False,
+                         allow_unicode=True)
+    content = f"---\n{yaml_str}---\n{body}"
+
+    with open(path, "w", encoding="utf-8") as f:
+        f.write(content)
+
+
+def update_frontmatter(path, updates, schema_type):
+    """Update specific frontmatter fields in an existing file.
+
+    Reads existing frontmatter, merges updates (overwriting on conflict),
+    validates, and writes back.
+
+    Args:
+        path: file path (must exist)
+        updates: dict of fields to add/update
+        schema_type: schema to validate against
+
+    Raises:
+        ValidationError: if merged data fails validation
+        FileNotFoundError: if file doesn't exist
+    """
+    data, body = read_frontmatter(path)
+    for key, value in updates.items():
+        if isinstance(value, dict) and isinstance(data.get(key), dict):
+            data[key].update(value)
+        else:
+            data[key] = value
+
+    apply_defaults(data, schema_type)
+    errors = validate(data, schema_type)
+    if errors:
+        raise ValidationError(
+            f"Frontmatter validation failed after update in {path}:\n"
+            + "\n".join(f"  - {e}" for e in errors))
+
+    yaml_str = yaml.dump(data, default_flow_style=False, sort_keys=False,
+                         allow_unicode=True)
+    content = f"---\n{yaml_str}---\n{body}"
+
+    with open(path, "w", encoding="utf-8") as f:
+        f.write(content)
