+++ scripts/utils/tc_parser.py
